// src/types/interfaces.ts
export interface infosProps {
  info: string;
  link?: string;
}

export interface competencesProps {
  name: string;
}

export interface coursesProps {
  name: string;
  course: string;
  date: string;
}

export interface professionalProps {
  company: string;
  office: string;
  date: string;
  description: string;
}